--- a/package.json
+++ b/package.json
@@ -121,6 +121,7 @@
     "graphql-tag": "^2.5.0",
     "isomorphic-unfetch": "^2.0.0",
     "load-script": "^1.0.0",
+    "lru-cache": "^4.1.1",
     "markdown-in-js": "^1.1.3",
     "mdast-react-render": "^1.1.0",
     "next": "^4.2.3",