cuts average SSR rendering time in half (for a really big mdast tree).

Rendering
/2018/02/26/render-stress-test (5x our longest article, [public 1x version](https://www.republik.ch/2018/01/15/demokratie-unter-irrationalen))
goes from a mean `542ms` to `282ms` (ab -n 1000 -c 4 against localhost).

**before**
```
$ ab -n 1000 -c 4 http://localhost:3010/2018/02/26/render-stress-test
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Completed 400 requests
Completed 500 requests
Completed 600 requests
Completed 700 requests
Completed 800 requests
Completed 900 requests
Completed 1000 requests
Finished 1000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            3010

Document Path:          /2018/02/26/render-stress-test
Document Length:        771012 bytes

Concurrency Level:      4
Time taken for tests:   135.570 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      771273000 bytes
HTML transferred:       771012000 bytes
Requests per second:    7.38 [#/sec] (mean)
Time per request:       542.280 [ms] (mean)
Time per request:       135.570 [ms] (mean, across all concurrent requests)
Transfer rate:          5555.78 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:   137  542 128.4    522    1625
Waiting:      136  432 130.3    420    1622
Total:        137  542 128.4    522    1626

Percentage of the requests served within a certain time (ms)
  50%    522
  66%    562
  75%    588
  80%    619
  90%    675
  95%    717
  98%    824
  99%    936
 100%   1626 (longest request)
```

**after**
```
$ ab -n 1000 -c 4 http://localhost:3010/2018/02/26/render-stress-test
This is ApacheBench, Version 2.3 <$Revision: 1807734 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient)
Completed 100 requests
Completed 200 requests
Completed 300 requests
Completed 400 requests
Completed 500 requests
Completed 600 requests
Completed 700 requests
Completed 800 requests
Completed 900 requests
Completed 1000 requests
Finished 1000 requests


Server Software:        
Server Hostname:        localhost
Server Port:            3010

Document Path:          /2018/02/26/render-stress-test
Document Length:        771023 bytes

Concurrency Level:      4
Time taken for tests:   70.519 seconds
Complete requests:      1000
Failed requests:        0
Total transferred:      771284000 bytes
HTML transferred:       771023000 bytes
Requests per second:    14.18 [#/sec] (mean)
Time per request:       282.074 [ms] (mean)
Time per request:       70.519 [ms] (mean, across all concurrent requests)
Transfer rate:          10680.98 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.1      0       1
Processing:    81  282 110.3    259    1512
Waiting:       81  246 105.1    227    1511
Total:         81  282 110.3    259    1513

Percentage of the requests served within a certain time (ms)
  50%    259
  66%    297
  75%    321
  80%    334
  90%    377
  95%    418
  98%    466
  99%    606
 100%   1513 (longest request)
```